api/requirements.txt
flask
tf-nightly
tensorflow
pillow
numpy
